requirements.txt
Flask
gunicorn
python-dotenv
redis